% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calanusthreshold-package.R
\docType{package}
\name{calanusthreshold-package}
\alias{calanusthreshold-package}
\alias{calanus}
\title{calanusthreshold-package}
\description{
Provides tools for producing threshold prey models for Calanus.
}
