% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{complement_species}
\alias{complement_species}
\title{Given a set of species, determine the other species}
\usage{
complement_species(species = "Calanus finmarchicus", known = known_species())
}
\arguments{
\item{species}{charcater, one or more species}
}
\value{
character vector, possibly empty, of the other species not identifed
}
\description{
Given a set of species, determine the other species
}
