% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{config_path}
\alias{config_path}
\title{Retrieve a configuration path}
\usage{
config_path(
  ...,
  root = rappdirs::user_config_dir(appname = "calanusthreshold")
)
}
\arguments{
\item{...}{character, file path segments passed to \code{file.path}}

\item{root}{character, the root configuration path}
}
\value{
file path
}
\description{
Retrieve a configuration path
}
