% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_subsample}
\alias{create_subsample}
\title{Given an input dataset, create a subsample and possibly anonymize ID and location}
\usage{
create_subsample(x, n = c(NA, 500)[2], anonymize = TRUE, noise = 0.2)
}
\arguments{
\item{x}{sf or tibble input dataset}

\item{n}{the number of subsamples, or NA to use all}

\item{anonymize}{logical, if TRUE anonymize the data}

\item{noise}{numeric, fraction of noise to add to each numeric variable (column)}
}
\value{
sf or tibble subset, possibly anonymized
}
\description{
Given an input dataset, create a subsample and possibly anonymize ID and location
}
