% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_project_path}
\alias{get_project_path}
\title{Retrieve the project path}
\usage{
get_project_path(filename = "project_path.txt", ...)
}
\arguments{
\item{filename}{character, the name of the filename with the config path}

\item{...}{other arguments for \code{config_path}}
}
\value{
the filepath to the project
}
\description{
Retrieve the project path
}
