% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{known_predictors}
\alias{known_predictors}
\title{The known predictor variables.}
\usage{
known_predictors()
}
\value{
character vector of allowed predictor variables
}
\description{
Year and station ID and location are disallowed.
Sea ice data (siconc, sithick) are too incomplete to train on.
}
