% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{known_species}
\alias{known_species}
\title{The known species groups}
\usage{
known_species(stage = FALSE)
}
\arguments{
\item{stage}{logical if FALSE drop the stage information}
}
\value{
character vector of known species names
}
\description{
The known species groups
}
