% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lump_vars}
\alias{lump_vars}
\title{Lump variables in a table together by summing}
\usage{
lump_vars(
  x = read_dataset(),
  vars = c("Calanus finmarchicus IV", "Calanus finmarchicus V",
    "Calanus finmarchicus VI", "Calanus hyperboreus IV", "Calanus hyperboreus V",
    "Calanus hyperboreus VI", "Calanus glacialis IV", "Calanus glacialis V",
    "Calanus glacialis VI"),
  selector = dplyr::all_of,
  newname = "lumped",
  na.rm = TRUE,
  drop = TRUE
)
}
\arguments{
\item{x}{tibble}

\item{vars}{character, one or more variables to lump}

\item{selector}{function the function used to select the variables,
see \code{tidyselect} helpers.}

\item{newname}{character, the name of the new (lumped) columns}

\item{na.rm}{logical, if TRUE lump after removing NA values}

\item{drop}{logical, if TRUE drop the \code{vars}}
}
\value{
tibble
}
\description{
Lump variables in a table together by summing
}
