% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Model the data}
\usage{
model(x, path = ".")
}
\arguments{
\item{x}{input}

\item{path}{character output path, if it doesn't exists the it is created}
}
\value{
model
}
\description{
Model the data
}
