% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_by_version}
\alias{model_by_version}
\title{Model the data for a given version configuration}
\usage{
model_by_version(
  cfg = read_config(system.file("exdata/cf.std.000.yaml", package =
    "calanusthreshold")),
  x = read_dataset(),
  save_model = FALSE,
  save_pred = FALSE,
  save_summary = FALSE,
  path = ".",
  verbose = interactive()
)
}
\arguments{
\item{cfg}{list, as per \code{\link{read_config}}}

\item{x}{input data set (as data.frame or tibble).  See \code{\link{read_dataset}}}

\item{save_model}{logical, if TRUE save the model, see \code{path}}

\item{save_pred}{logical, if TRUE save the prediction (made with testing data split), see \code{path}}

\item{save_summary}{logical, if TRUE save the summary (made with testing data split), see \code{path}}

\item{path}{character, output path, if it doesn't exists the it is created. Subdirectories
are created for saving model, prediction, and summary}

\item{verbose}{logical, if TRUE output helpful messages}
}
\value{
\code{\link[parsnip]{model_fit}} object or NULL
This gets returned when there is an issue
}
\description{
Model the data for a given version configuration
}
