% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_predvars}
\alias{model_predvars}
\title{Extract the names of the predictor variables}
\usage{
model_predvars(x)
}
\arguments{
\item{x}{model inheriting from \code{model_fit}}
}
\value{
charcater vector of predictor names
}
\description{
Extract the names of the predictor variables
}
