% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_vars}
\alias{model_vars}
\title{Extract the names of the predictor and predicted variables}
\usage{
model_vars(x)
}
\arguments{
\item{x}{model inheriting from \code{model_fit}}
}
\value{
list of \code{predicted} and \code{predictors} names
}
\description{
Extract the names of the predictor and predicted variables
}
