% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Plot prediction rasters (patch and probability)}
\usage{
plot_prediction(x, zlim = list("fit", c(0, 1))[[1]])
}
\arguments{
\item{x}{\code{stars} object with two attributes}

\item{zlim}{2 element vector of limits for probability, or 
"fit" to scale to the range of prob values}
}
\value{
2 element list of ggplot2 objects ("patch", "prob")
}
\description{
Plot prediction rasters (patch and probability)
}
