% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_calanus}
\alias{predict_calanus}
\title{Predict using a model}
\usage{
predict_calanus(model, predictors, include_prob = TRUE, ...)
}
\arguments{
\item{model}{either a 'model_fit' object or a file description to one}

\item{predictors}{a table (data.frame or tibble) or raster (terra or stars)}

\item{include_prob}{logical, if TRUE include probabilities of occurrence}

\item{...}{arguments for \code{predict}}
}
\value{
prediction as table (tibble) or stars raster
}
\description{
Predict using a model
}
