% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_raster}
\alias{predict_raster}
\title{Predict using stars class raster}
\usage{
predict_raster(
  model,
  predictors = read_raster(),
  include_prob = TRUE,
  month = NA,
  drop = TRUE,
  ...
)
}
\arguments{
\item{model}{either a 'model_fit' object or a file description to one}

\item{predictors}{a table (data.frame or tibble) or raster (terra or stars)}

\item{include_prob}{logical, if TRUE include probabilities of occurrence}

\item{month}{numeric or NA, if not NA then add a layer with value equal to this month number}

\item{drop}{logical, if TRUE drop unused layers in the predictors}

\item{...}{arguments passed to \code{predict}}
}
\value{
stars object with one or two layers (see \code{include_prob})
}
\description{
Predict using stars class raster
}
