% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_table}
\alias{predict_table}
\title{Predict using tibble class object}
\usage{
predict_table(model, predictors, include_prob = TRUE, ...)
}
\arguments{
\item{model}{either a 'model_fit' object or a file description to one}

\item{predictors}{a table (data.frame or tibble) or raster (terra or stars)}

\item{include_prob}{logical, if TRUE include probabilities of occurrence}

\item{...}{arguments passed to \code{predict}}
}
\value{
tibble of predictions, possibly with prob.  In the event that the
  predictors have no complete cases (at least one NA in each row) then 
  the prediction returned is NA everywhere with a warning
}
\description{
Predict using tibble class object
}
