% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{prep_dataset}
\alias{prep_dataset}
\title{Prep a dataset by removing variables, remove incolplete cases,
 aggregating specified columns, and applying labeling to the aggregate column}
\usage{
prep_dataset(
  x = read_dataset(form = "tibble"),
  drop_var = c("Calanus glacialis", "Calanus hyperboreus", "geom", "longitude",
    "latitude", "station", "year", "siconc", "sithick", "month"),
  drop_fun = dplyr::starts_with,
  complete_cases_only = TRUE,
  lump_var = "Calanus finmarchicus",
  lump_fun = dplyr::starts_with,
  newname = "patch",
  threshold = 10000
)
}
\arguments{
\item{x}{tibble of input data}

\item{drop_var}{character, the variables to deselect}

\item{drop_fun}{function, such as \code{\link[dplyr]{starts_with}}}

\item{complete_cases_only}{logical, if TRUE drop incomplete records (after deselection)}

\item{lump_var}{character character, vector of variables to sum into a new column}

\item{lump_fun}{function, such as \code{\link[dplyr]{starts_with}}}

\item{newname}{character the name of the new aggregated colum}

\item{threshold}{numeric, the threshold(s) used to define patches}
}
\value{
a tibble
}
\description{
Prep a dataset by removing variables, remove incolplete cases,
 aggregating specified columns, and applying labeling to the aggregate column
}
