% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{project_path}
\alias{project_path}
\title{Retrieve project data path}
\usage{
project_path(..., root = get_project_path())
}
\arguments{
\item{root}{character, the root project path}

\item{filename}{character, the name of the filename with the project path}
}
\value{
a fully qualified path
}
\description{
Retrieve project data path
}
