% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_config}
\alias{read_config}
\title{Read configuration}
\usage{
read_config(filename)
}
\arguments{
\item{filename}{the file to read}
}
\value{
a list of configuration values
}
\description{
Read configuration
}
