% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_dataset}
\alias{read_dataset}
\title{Read an input file source}
\usage{
read_dataset(
  filename = system.file("exdata/fake_data.csv.gz", package = "calanusthreshold"),
  drop_vars = c("station", "latitude", "longitude", "siconc", "sithick", "geometry"),
  na.rm = TRUE
)
}
\arguments{
\item{filename}{character, the file to read. The default is to read a dummy example data file.}

\item{drop_vars}{character or NULL, variables (columns) to drop from input}

\item{na.rm}{logical, if TRUE drop any cases (rows) with one or more missing values}
}
\value{
tibble or simple feature
}
\description{
Read an input file source
}
