% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_model}
\alias{read_model}
\title{Read a model from a file}
\usage{
read_model(
  filename = system.file("exdata/model.Rdata", package = "calanusthreshold")
)
}
\arguments{
\item{filename}{character, the name of the file to read}
}
\value{
the model
}
\description{
Read a model from a file
}
