% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_pred}
\alias{read_pred}
\title{Read one or more prediction files}
\usage{
read_pred(
  filename = system.file("exdata/prediction.csv.gz", package = "calanusthreshold")
)
}
\arguments{
\item{filename}{charcater, one or more filenames}
}
\value{
tibble (possibly with a 'version' column)
}
\description{
If multiple files are provided then a "version" column is prepended
}
