% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_raster}
\alias{read_raster}
\title{Read a raster}
\usage{
read_raster(
  filename = system.file("exdata/predictors.tif", package = "calanusthreshold"),
  form = c("raster", "tibble", "sf")[1]
)
}
\arguments{
\item{filename}{character, the name of the file to read}

\item{form}{characater, one of 'raster', 'sf' or 'tibble'.  If the latter then
coordinate columns (x, y) followed by one column per variable.}
}
\value{
raster or tibble object
}
\description{
Read a raster
}
