% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_model}
\alias{save_model}
\title{Save a model to a file}
\usage{
save_model(x, filename = "calanusthreshold_model.Rdata", overwrite = FALSE)
}
\arguments{
\item{x}{model structure}

\item{filename}{character, the name of the file to save.}

\item{overwrite}{logical, overwrite existing files? It is an error to try to without
setting this argument to TRUE}
}
\value{
the input model
}
\description{
Save a model to a file
}
