% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{set_project_path}
\alias{set_project_path}
\title{Set the data path as a configuration}
\usage{
set_project_path(path = ".", ...)
}
\arguments{
\item{path}{character, the path to the project data}

\item{...}{other arguments for \code{config_path}}
}
\value{
logical with TRUE for success
}
\description{
Set the data path as a configuration
}
