% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_config}
\alias{write_config}
\title{Write configuration}
\usage{
write_config(x, filename = "config.yaml")
}
\arguments{
\item{x}{configuration (a list)}

\item{filename}{the file to write}
}
\value{
a list of configuration values
}
\description{
Write configuration
}
