% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_dataset}
\alias{write_dataset}
\title{Write a dataset}
\usage{
write_dataset(x, filename = "dataset.csv.gz")
}
\arguments{
\item{x}{sf or tibble input-style dataset}

\item{filename}{character, the file to write to}
}
\value{
the provided sf or tibble
}
\description{
Write a dataset
}
