% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_pred}
\alias{write_pred}
\title{Write a prediction file}
\usage{
write_pred(x, filename)
}
\arguments{
\item{x}{table of predictions}

\item{filename}{character, the filename to write to}
}
\value{
the inout, \code{x}
}
\description{
Write a prediction file
}
