% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_raster}
\alias{write_raster}
\title{Write a raster}
\usage{
write_raster(x, filename = "raster.tif", overwrite = TRUE)
}
\arguments{
\item{x}{raster object}

\item{filename}{character, the name of the file to write}

\item{overwrite}{logical, if TRUE allow overwrite of an existing file}
}
\value{
stars raster object
}
\description{
Write a raster
}
