% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as_patch}
\alias{as_patch}
\title{Convert abundance to patch density}
\usage{
as_patch(
  x,
  threshold = list(10000, c(2500, 5000, 7500, 10000))[[1]],
  form = c("index", "factor")[2]
)
}
\arguments{
\item{x}{a vector of abundances}

\item{threshold}{numeric boundaries defining levels of patchiness from low to high.
Values in \code{x} below the lowest threshold are considered "no-patch"}

\item{form}{character, one of 'index' of 'factor' to control the output type
in either case NAs are propagated.}
}
\value{
numeric indices into threshold, possibly as factors.  Zero means no patch.
}
\description{
Convert abundance to patch density
}
