% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{prep_dataset}
\alias{prep_dataset}
\title{Prep a dataset by removing variables, remove incomplete cases,
 aggregating specified columns, and applying threshold/labeling to the aggregated column}
\usage{
prep_dataset(
  x = read_dataset(),
  complete_cases_only = TRUE,
  lump_var = "Calanus finmarchicus",
  lump_fun = dplyr::starts_with,
  drop_var = c("geometry", "longitude", "latitude", "station", "year", "siconc",
    "sithick", complement_species(lump_var)),
  drop_fun = dplyr::starts_with,
  log_var = c("bathymetry", "chlor_a"),
  newname = "patch",
  threshold = 10000
)
}
\arguments{
\item{x}{tibble of input data}

\item{complete_cases_only}{logical, if TRUE drop incomplete records (after deselection)}

\item{lump_var}{character character, vector of variables to sum into a new column}

\item{lump_fun}{function, such as \code{\link[dplyr]{starts_with}}}

\item{drop_var}{character, the variables to deselect}

\item{drop_fun}{function, such as \code{\link[dplyr]{starts_with}}}

\item{log_var}{character, variables to log scale (base 10), match with \code{\link[dplyr]{any_of}}}

\item{newname}{character the name of the new aggregated colum}

\item{threshold}{numeric, the threshold(s) used to define patches}
}
\value{
a tibble cleaned and transformed for downstream processing
}
\description{
Prep a dataset by removing variables, remove incomplete cases,
 aggregating specified columns, and applying threshold/labeling to the aggregated column
}
