% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{project_path}
\alias{project_path}
\title{Retrieve project data path}
\usage{
project_path(..., root = get_project_path())
}
\arguments{
\item{...}{character, file path segments for file/directories relative to \code{root}
See \code{\link[base]{file.path}}}

\item{root}{character, the root project path}
}
\value{
a fully qualified path
}
\description{
Retrieve project data path
}
