% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_dataset}
\alias{read_dataset}
\title{Read an input file source}
\usage{
read_dataset(
  filename = system.file("exdata/example_input.csv.gz", package = "calanusthreshold"),
  form = c("tibble", "sf")[1],
  crs = 4326
)
}
\arguments{
\item{filename}{character, the file to read. The default is to read a dummy example data file.}

\item{form}{character, either 'tibble' (default) or 'sf'}

\item{crs}{character or numeric, if \code{form} is 'sf' then pass this to
\code{\link[sf]{st_as_sf}}}
}
\value{
tibble or simple feature
}
\description{
Read an input file source
}
